djangoapp/requirements.txt
Django==5.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.1
coreapi==2.3.3
PyYAML==6.0.1
drf-spectacular==0.27.2
django-filter==24.2
Pint==0.24.3




